+++ README.md
@@ -79,9 +79,9 @@
   - [x] Enrollment 검증 테스트 도입
 - [x] 강의 진행 상태 / 모집 상태 분리
 - [x] 강의 커버 이미지 다수 등록
-- [ ] 수강 신청 변경
-  - [ ] 수강 신청 상태 추가 (미승인/승인/취소)
+- [x] 수강 신청 변경
+  - [x] 수강 신청 상태 추가 (미승인/승인/취소)
     - [x] Enrollment 추가
-    - [ ] Enrollments 리팩토링
-  - [ ] 수강 신청 승인
-  - [ ] 수강 신청 취소
+    - [x] Enrollments 리팩토링
+  - [x] 수강 신청 승인
+  - [x] 수강 신청 취소
